third_party/WebKit/LayoutTests/inspector/progress-bar-expected.txt

Tests inspector's composite progress bar.


Running: testOneSubProgress
Testing CompositeProgress with a single subprogress:
progress: `undefined' 0 out of 1 done.
progress: `cuckooing' 0 out of 1 done.
progress: `cuckooing' 0.1 out of 1 done.
progress: `meowing' 0.2 out of 1 done.
Progress composite dispatched Done event.
progress indicator: done
Progress subProgress dispatched Done event.
progress: `meowing' 1 out of 1 done.

Running: testMultipleSubProgresses
Testing CompositeProgress with multiple subprogresses:
progress: `undefined' 0 out of 1 done.
progress: `cuckooing' 0.05 out of 1 done.
progress: `cuckooing' 0.05 out of 1 done.
progress: `barking' 0.275 out of 1 done.
progress: `meowing' 0.5 out of 1 done.
Progress subProgress2 dispatched Done event.
progress: `meowing' 0.875 out of 1 done.
Progress composite dispatched Done event.
progress indicator: done
Progress subProgress1 dispatched Done event.
progress: `meowing' 1 out of 1 done.

Running: testCancel
Testing isCanceled:

Running: testNested
Testing nested subprogresses:
progress: `undefined' 0 out of 1 done.
progress: `undefined' 0 out of 1 done.
progress: `undefined' 0.1 out of 1 done.
progress: `meowing' 0.5 out of 1 done.
Progress composite1 dispatched Done event.
Progress composite0 dispatched Done event.
progress indicator: done
Progress subProgress01 dispatched Done event.
Progress subProgress11 dispatched Done event.
progress: `meowing' 1 out of 1 done.
